one time connect and receive data from esp
but when send data to esp connection fail and not fix yet:(

+++ Core/Src/motor_Control/motor_Control.h
@@ -48,7 +48,7 @@ typedef struct
 	Encoder_State_Enum Pre_Encoder_State;
 	_Bool Stuck_Finger;
 	uint8_t speed;
-	uint32_t Current;
+	uint16_t Current;
 	volatile _Bool SignalA;
 	volatile _Bool SignalB;
 }Finger_Struct;
